feat(ui): add URL-based routing for session tabs
Refactor session page to use URL-based routing for chat, files, and
events sections. Each tab now has its own URL:
- /agents/{id}/sessions/{id}/chat
- /agents/{id}/sessions/{id}/files
- /agents/{id}/sessions/{id}/events

This enables better navigation, bookmarking, and sharing of specific
session views.

Changes:
- Create SessionProvider context for shared state across tab pages
- Add layout.tsx with shared header and tab navigation
- Split page.tsx into separate chat, files, and events pages
- Base session URL redirects to /chat by default
